refactor(core): extract createOverlayEffects
Consolidate three small infra effects in init() into one named factory:

- detectedElement liveness poll (clears when DOM removes the node)
- frozenElements WAAPI animation freeze
- freezeReactUpdates while activated

The factory wires them into the active Solid root so they tear down
when init() disposes; each effect owns its own onCleanup. Drops 3
now-unused imports (freezeAnimations, freezeUpdates,
BOUNDS_RECALC_INTERVAL_MS).

index.tsx: 2921 -> 2897 (-24).
